move division rules to their own file
