+++ backend/tests/divisors.py
@@ -2,8 +2,8 @@ import math
 import unittest
 
 from voting import apportion1d
-from voting import dhondt_gen, sainte_lague_gen, nordic_sainte_lague_gen, \
-                   imperiali_gen, danish_gen, huntington_hill_gen
+from division_rules import dhondt_gen, sainte_lague_gen, \
+    nordic_sainte_lague_gen, imperiali_gen, danish_gen, huntington_hill_gen
 
 class TestDivisors(unittest.TestCase):
